chore: move load-commands file to infrastructure/discrod
