+++ src/infrastructure/discord/load-commands.js
